+++ cli/tests/runner.rs
@@ -14,6 +14,7 @@ mod test_acls;
 mod test_advance_bookmarks;
 mod test_alias;
 mod test_backout_command;
+mod test_bisect_command;
 mod test_bookmark_command;
 mod test_builtin_aliases;
 mod test_commit_command;